Move Substitution into subst.rs
